Testing : favor composition over inheritance
